Move types around again, make sure eslint checks parent package.json deps
