+++ .vscode/settings.json
@@ -9,5 +9,6 @@
     {
       "pattern": "types"
     }
-  ]
+  ],
+  "typescript.tsdk": "node_modules\\typescript\\lib"
 }
