+++ rabbitmq/extension/package.json
@@ -12,7 +12,6 @@
     "amqplib": "^0.7.0"
   },
   "devDependencies": {
-    "@esamarathon/mq-events": "github:esamarathon/mq-events",
     "@types/amqp-connection-manager": "^2.0.10",
     "@types/amqplib": "^0.5.17"
   }
