+++ types/OBS.d.ts
@@ -1,0 +1,7 @@
+export namespace OBS {
+  interface Config {
+    enable: boolean;
+    address: string;
+    password: string;
+  }
+}
