+++ types/XKeys.d.ts
@@ -1,0 +1,5 @@
+export namespace XKeys {
+  interface Config {
+    enable: boolean;
+  }
+}
