+++ types/index.d.ts
@@ -1,5 +1,8 @@
 export * from './Asset';
 export * from './MediaBox';
+export * from './OBS';
 export * from './RabbitMQ';
 export * from './Restream';
 export * from './Tracker';
+export * from './X32';
+export * from './XKeys';
