+++ x32/extension/src/index.ts
@@ -1,15 +1,10 @@
 import type { NodeCG } from 'nodecg/types/server';
 import osc from 'osc';
-
-interface Config {
-  enable: boolean;
-  ip: string;
-  localPort: number;
-}
+import { X32 as X32Types } from '../../../types';
 
 class X32 {
   private nodecg: NodeCG;
-  private config: Config;
+  private config: X32Types.Config;
   conn: osc.UDPPort | undefined;
   faders: { [k: string]: number } = {};
   fadersExpected: { [k: string]: {
@@ -17,7 +12,7 @@ class X32 {
   } } = {};
   private fadersInterval: { [k: string]: NodeJS.Timeout } = {};
 
-  constructor(nodecg: NodeCG, config: Config) {
+  constructor(nodecg: NodeCG, config: X32Types.Config) {
     this.nodecg = nodecg;
     this.config = config;
